requirements.txt
fastapi
uvicorn[standard]
pdfplumber
anthropic
sqlalchemy
psycopg2-binary
python-multipart
aiofiles
fpdf2
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
stripe
genanki
